Create FeatureCollections for relations
This is a simplistic, but general way of creating geometries.
Utilise queuing to process the ways and relations once all the required
elements have been stored in the database. This fixes the issue with
queries not returning the required data. This was happening because the
save has only completed once the callback has been excuted.