Fixes to way processing
Fix handling of options.useOriginalID in several places.
Move the async.queue callback to the query callback, as before it was
querying the database uncontrollably.
Set way concurrency to 5000 (this is a value that worked for me), as 1
was too slow when processing large data files. Because of the above
callback move, this can be treated as the maximum number of concurrent
database queries.